Move `markCandidates` into its own module
